pyproject.toml
 [tool.poetry]
name = "hccc-fastapi-sso"
version = "0.1.0"
description = "FastAPI project for HCCC with SSO"
authors = ["Jay Singh <jsingh@hccc.edu>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
httpx = "^0.26.0"
orjson = "^3.9.13"
email-validator = "^2.1.0.post1"
gunicorn = "^21.2.0"
psutil = "^5.9.8"
trio = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "0.25.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.2.1"
mypy = "^1.8.0"
pytest-postgresql = "^5.0.0"
pytest-env = "^1.1.3"
pytest-xdist = "^3.5.0"
pytest-sugar = "^1.0.0"
faker = "^22.6.0"
types-psutil = "^6.1.0.20241221"
pytest-mock = "^3.14.0"
types-passlib = "^1.7.7.20241221"
types-python-jose = "^3.3.4.20240106"

[tool.poetry.group.e2e.dependencies]
playwright = "^1.50.0"
pytest-playwright = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

# Pydantic specific mypy configurations
[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html:coverage-reports/html --cov-report=xml:coverage-reports/coverage.xml -n auto"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "strict"
markers = [
    "asyncio: mark test as async",
    "e2e: mark test as end-to-end test",
    "unit: mark test as unit test",
    "integration: mark test as integration test",
    "slow: mark test as slow running",
    "db: mark test as requiring database"
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
env = [
    "POSTGRES_SERVER=test_db",
    "POSTGRES_USER=postgres",
    "POSTGRES_PASSWORD=change-in-production",
    "POSTGRES_DB=test_user_management",
    "POSTGRES_PORT=5432",
    "DEBUG=true"
]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs/
'''
